requirements.txt
flask >= 3.0.0
Pillow >= 10.2
pdoc >= 14.4
